+++ app/src/main/java/eu/kanade/presentation/more/settings/widget/ListPreferenceWidget.kt
@@ -35,6 +35,7 @@ fun <T> ListPreferenceWidget(
     icon: ImageVector?,
     entries: Map<out T, String>,
     onValueChange: (T) -> Unit,
+    enabled: Boolean = true,
 ) {
     var isDialogShown by remember { mutableStateOf(false) }
 
@@ -42,7 +43,8 @@ fun <T> ListPreferenceWidget(
         title = title,
         subtitle = subtitle,
         icon = icon,
-        onPreferenceClick = { isDialogShown = true },
+        onPreferenceClick = { if (enabled) isDialogShown = true },
+        enabled = enabled,
     )
 
     if (isDialogShown) {